+++ client/components/frequencies/frequencies.ts
@@ -59,11 +59,14 @@ namespace frequencies {
 				function($scope: any, $state: any, $stateParams: any, $http: angular.IHttpService) {
 					
 					$scope.deleteFrequency = function (data: any)  {
-						$http.delete(`${url}/${data.id}`).then((resp) => {
-							if (resp.data['success']) {
-								$state.go($state.current, {}, {reload: true});
-							}
-						}); 
+						let result: boolean = confirm("Are you sure?");
+						if (result) {
+							$http.delete(`${url}/${data.id}`).then((resp) => {
+								if (resp.data['success']) {
+									$state.go($state.current, {}, {reload: true});
+								}
+							});
+						} 
 					};
 					
 					$http.get(`${url}/_find`).then((resp) => {
